Refactor MyRenderHelper and Helper methods to improve parameter types for better compatibility

+++ src/ipstubs/java/qouteall/imm_ptl/core/render/MyRenderHelper.java
@@ -6,8 +6,8 @@ public class MyRenderHelper {
     public static void drawFramebuffer(
         RenderTarget target,
         boolean flipY, boolean flipX,
-        double left, double right,
-        double bottom, double top
+        float left, double right,
+        float bottom, double top
     ) {
     }
 }
